Move existing fixture into new `test/fixtures` dir
