+++ loader/test/fixtures/cvs.smart.fixtures.js
