+++ loader/test/fixtures/nock/washington_doh_api_should_successfully_format_results.json
@@ -10,7 +10,7 @@
                     "state": "PR",
                     "paging": {
                         "pageNum": 1,
-                        "pageSize": 100
+                        "pageSize": 200
                     }
                 }
             }
